Fix 'blog' template for courses and workshops.
Use a tag listing method that more closely resembles the one used in the default templates that come with the Quark theme. This means that when we click a tag and see only its posts, the tag listing still shows *all* tags and not only the tags of the posts that are now being shown.
Also disable the justified text when listing the tags and fix some of the links that didn't scroll automatically to the #body-wrapper section of the page.